Move converter to another file
